feat: login page integrated with backend
First of all, I would like to say that the code is not "elegant" and/or optimized and will undoubtedly be refactored and modified in the next commits. That being said, the login screen is now working and the users can provide their Instagram usernames and passwords in a kivy window and watch Selenium do the rest.